refactor: address review feedback on proposal query endpoints

+++ src/backend/src/dto/proposal.rs
@@ -21,7 +21,7 @@ pub struct ListProjectProposalsRequest {
     pub project_id: String,
     pub status_filter: Option<Vec<ProposalStatusFilter>>,
     pub after: Option<String>,
-    pub limit: Option<u32>,
+    pub limit: Option<u64>,
 }
 
 #[derive(Debug, Clone, CandidType, Deserialize)]
